Testing with docker
This patch contains a simple enough Dockerfile that tries to build a
'core' and a 'numerical' profile. The Dockerfile should be broken into
two docker files so there base system can be referenced for other Docker
projects.

Note, parmetis is currently not building in this container. There may be
other broken applications as this has not successfully run to completion.